job-agent-mvp/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.36.0
pydantic>=2.8.0
python-dotenv>=1.0.1
openai>=1.40.0
httpx>=0.27.0
